Changed monkey images (except for poster) back to png, changed background to css gradient
The monkeys needed to be in a format that allowed transparency (which jpg does not allow). I used a gradient creator to make a somewhat similar background to the picture background that I originally had. Reduced the size of the image I was using as a poster and then compressed that.